Daily progress
Finished the pattern converter module to take BCD and convert it to the pattern the LCD Diplay Data RAM needs to make a character.

Started on the interface inside the top level module to interface from the combinational logic to the sequential logic of the lcdController.

Instantiated some of the combinaitonal logic modules in the top level module and started to wire them together.

Deleted Quartus auto generated back up files.

renamed files and variables to lowercase "lcd..." for consistency.